requirements.txt
aiohttp==1.0.5
async-timeout==1.2.0
cffi==1.10.0
chardet==2.3.0
discord.py==0.16.7
multidict==2.1.4
opuslib==1.1.0
pycparser==2.17
PyNaCl==1.0.1
six==1.10.0
virtualenv==15.1.0
websockets==3.2
youtube-dl==2017.3.22
